Make the check stemcell version task generic
With this PR, the check stemcell version is made generic
so it can accommodate develop branch. As this will be used
to make comparison between either develop or release-candidate branch
and main branch.
