--- tasks/check-stemcell-versions-for-ship-it/task
@@ -1,11 +1,0 @@
-#!/bin/bash
-set -eu
-
-function main() {
-  root_dir="$PWD"
-  pushd "$(dirname $0)"
-    go run main.go "${root_dir}"
-  popd
-}
-
-main
