--- tasks/check-stemcell-versions-for-ship-it/task.yml
@@ -1,15 +1,0 @@
----
-platform: linux
-
-image_resource:
-  type: docker-image
-  source:
-    repository: cloudfoundry/relint-base
-
-inputs:
-- name: runtime-ci
-- name: cf-deployment-release-candidate
-- name: cf-deployment-main
-
-run:
-  path: runtime-ci/tasks/check-stemcell-versions-for-ship-it/task
